build: configure dependecies to allow WASM compilation
