+++ ls/Cargo.toml
@@ -9,21 +9,29 @@ license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
 
+[lib]
+name = "yara_x_ls"
+crate-type = ["cdylib", "rlib"]
+
 [[bin]]
 name = "yr-ls"
 path = "src/main.rs"
 
 [dependencies]
-async-lsp = { version = "0.2.2", default-features=false, features=["omni-trait", "tokio", "client-monitor", "stdio"]}
 bitflags = { workspace = true }
 futures = "0.3.31"
-tokio = { version = "1.48.0", features = ["full"] }
-tokio-util = { version = "0.7.17", features = ["compat"] }
-tower = { version = "0.5.2" , features = ["full"]}
 yara-x-parser = { workspace = true }
 
+[target.'cfg(not(any(target_arch = "wasm32", target_arch = "wasm64")))'.dependencies]
+tokio = { version = "1.48.0", features = ["full"] }
+tokio-util = { version = "0.7.17", features = ["compat"] }
+async-lsp = {version = "0.2.2", default-features=false, features=["omni-trait", "tokio", "stdio"]}
+tower = { version = "0.5.2" , features = ["full"]}
+
+[target.'cfg(any(target_arch = "wasm32", target_arch = "wasm64"))'.dependencies]
+async-lsp = {version = "0.2.2", default-features=false, features=["omni-trait"]}
+tower = { version = "0.5.2" }
 
 [dev-dependencies]
 serde = { workspace = true }
 serde_json = { workspace = true }
-
