+++ ls/src/lib.rs
@@ -1,0 +1,33 @@
+use async_lsp::{
+    concurrency::ConcurrencyLayer, panic::CatchUnwindLayer,
+    server::LifecycleLayer,
+};
+use futures::{AsyncRead, AsyncWrite};
+use tower::ServiceBuilder;
+
+use crate::server::ServerState;
+
+mod features;
+mod server;
+#[cfg(test)]
+mod tests;
+mod utils;
+
+/// Starts the Language Server Main Loop with provided streams.
+///
+/// Provided streams must implement [`futures::AsyncRead`] and
+/// [`futures::AsyncWrite`] traits.
+pub async fn serve(
+    input: impl AsyncRead,
+    output: impl AsyncWrite,
+) -> Result<(), async_lsp::Error> {
+    let (server, _) = async_lsp::MainLoop::new_server(|client| {
+        ServiceBuilder::new()
+            .layer(LifecycleLayer::default())
+            .layer(CatchUnwindLayer::default())
+            .layer(ConcurrencyLayer::default())
+            .service(ServerState::new_router(client))
+    });
+
+    server.run_buffered(input, output).await
+}
